+++ pages/account.js
@@ -1,7 +1,7 @@
 import { Text, View, Image, ScrollView, TouchableOpacity, Linking } from "react-native";
 import { useEffect, useState } from "react";
 import styles from "../styles/account/style";
-import BottomNavBar from "../global_components/bottom_navbar";
+import BottomNavBar from "../components/global/bottom_navbar";
 import getEnvVars from "../variables";
 import * as SecureStore from 'expo-secure-store';
 import { useWebSocket } from "../scripts/websocket_handler";
